[refactor] 등수 구하는 책임 위임
